pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "syft-accounting-sdk"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = ["requests==2.32.3", "pydantic==2.11.4", "click==8.1.3"]

[project.scripts]
accounting = "syft_accounting_sdk.cli:accounting"
accounting-admin = "syft_accounting_sdk.cli:accounting_admin"

[project.optional-dependencies]
dev = ["jupyter==1.1.1", "python-dotenv==1.1.0"]
cli = ["click==8.1.3", "python-dotenv==1.1.0", "colorama==0.4.6"]

[tool.hatch.build.targets.wheel]
packages = ["src/syft_accounting_sdk"]